requirements.txt
Django==3.1.1
django-allauth==0.42.0
django-cors-headers==3.2.1
django-filter==2.4.0
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-url-filter==0.3.15
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
psycopg2-binary==2.8.6
enum-compat==0.0.3
hyperlink==20.0.1
idna==2.9
incremental==17.5.0
msgpack==0.6.2
oauthlib==3.1.0
Pillow==7.0.0
pyOpenSSL==19.1.0
python-http-client==3.2.7
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
urllib3==1.25.9
zope.interface==5.1.0
pyotp==2.4.1
kombu==5.0.2
boto3
pdfkit
wkhtmltopdf
wheel
xlrd
gunicorn==20.0.4
